tidy up functions
